+++ src/lib.rs
@@ -90,8 +90,8 @@ let html = parser.words_to_html(&repetitions, 2.0, true):
 */
 
 // Uncomment this if you use nightly and want to run benchmarks
-//#![feature(test)]
-//mod bench;
+#![feature(test)]
+mod bench;
 mod word;
 mod parser;